feat: refactor HTML structure and styles, enhance user authentication flow, and implement cookie management
