+++ admin.html
@@ -9,99 +9,13 @@
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
   <!-- Bootstrap 5 -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
-
-  <!-- Light / Dark Theme CSS -->
-  <style>
-    :root {
-      --bg: #f4f6f9;
-      --text: #222;
-      --sidebar-bg: #d6d5f5;
-      --sidebar-text: #333;
-      --sidebar-active: #bdbaf4;
-      --card-bg: #ffffff;
-    }
-
-    body.dark {
-      --bg: #1e1e2f;
-      --text: #eee;
-      --sidebar-bg: #2f2f4f;
-      --sidebar-text: #ccc;
-      --sidebar-active: #49497c;
-      --card-bg: #2b2b3d;
-    }
-
-    body {
-      font-family: 'Poppins', sans-serif;
-      background-color: var(--bg);
-      color: var(--text);
-      margin: 0;
-    }
-
-    .sidebar {
-      height: 100vh;
-      background-color: var(--sidebar-bg);
-      padding: 1.5rem;
-      color: var(--sidebar-text);
-    }
-
-    .sidebar h4 {
-      font-weight: 600;
-      color: var(--text);
-    }
-
-    .sidebar a {
-      display: block;
-      color: var(--sidebar-text);
-      margin: 0.75rem 0;
-      text-decoration: none;
-      padding: 0.5rem 1rem;
-      border-radius: 8px;
-    }
-
-    .sidebar a:hover,
-    .sidebar a.active {
-      background-color: var(--sidebar-active);
-      color: var(--text);
-    }
-
-    .main {
-      padding: 2rem;
-    }
-
-    .section {
-      display: none;
-    }
-
-    .section.active {
-      display: block;
-    }
-
-    .table, .table thead {
-      background-color: var(--card-bg);
-      color: var(--text);
-    }
-
-    .theme-toggle {
-      margin-top: 1rem;
-      text-align: center;
-    }
-
-    .btn-theme {
-      font-size: 0.9rem;
-      background: transparent;
-      border: 1px solid #888;
-      border-radius: 6px;
-      padding: 4px 12px;
-      color: inherit;
-      cursor: pointer;
-    }
-  </style>
+    <link rel="stylesheet" href="css/admin.css">    
 </head>
 
 <body>
   <div class="container-fluid">
     <div class="row">
-      <!-- Sidebar -->
+   
       <nav class="col-md-3 col-lg-2 sidebar shadow-sm">
         <h4>🌤️ Admin Panel</h4>
         <a href="#" class="nav-link active" data-target="dashboard">🏠 Tổng quan</a>
@@ -110,15 +24,10 @@
 
         <hr />
         <div>
-          <p class="mb-1">👤 <span id="admin-email">Đang tải...</span></p>
-          <button id="logout-btn" class="btn btn-sm btn-outline-dark mt-2">Đăng xuất</button>
-        </div>
-        <div class="theme-toggle">
-          <button id="toggle-theme" class="btn-theme">🌞 / 🌙</button>
+         <a id="logout" class="btn btn-sm btn-outline-dark mt-2" href="#">Đăng xuất</a>
         </div>
       </nav>
-
-      <!-- Main Content -->
+    
       <main class="col-md-9 col-lg-10 main">
         <div id="dashboard" class="section active">
           <h2 class="mb-3">📍 Top địa điểm tra cứu</h2>
@@ -164,17 +73,8 @@
         document.getElementById(link.dataset.target).classList.add("active");
       });
     });
-
-    // Giao diện sáng / tối
-    const root = document.body;
-    const themeBtn = document.getElementById("toggle-theme");
-    const savedTheme = localStorage.getItem("theme");
-    if (savedTheme === "dark") root.classList.add("dark");
-
-    themeBtn.addEventListener("click", () => {
-      root.classList.toggle("dark");
-      localStorage.setItem("theme", root.classList.contains("dark") ? "dark" : "light");
-    });
+    
   </script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
