+++ css/account.css
@@ -1,0 +1,53 @@
+body {
+  background-color: #e0f7fa;
+  font-family: "Poppins", sans-serif;
+}
+
+.navbar {
+  background-color: #b2ebf2 !important;
+}
+
+.card-profile {
+  background-color: #ffffff;
+  border-radius: 16px;
+  padding: 2.5rem;
+  max-width: 600px;
+  margin: 60px auto;
+  box-shadow: 0 8px 20px rgba(0, 172, 193, 0.1);
+}
+
+h4 {
+  color: #00acc1;
+  font-weight: 600;
+  text-align: center;
+  margin-bottom: 1.5rem;
+}
+
+.form-control {
+  border-radius: 10px;
+}
+
+.form-label {
+  font-weight: 500;
+}
+
+.btn-save {
+  background-color: #00acc1;
+  border: none;
+  color: #fff;
+  font-weight: 500;
+  border-radius: 10px;
+}
+
+.btn-save:hover {
+  background-color: #00838f;
+}
+
+#previewAvatar {
+  width: 400px;
+  height: 400px;
+  border-radius: 50%;
+  border: 3px solid #00bcd4;
+  object-fit: cover;
+  box-shadow: 0 4px 8px rgba(0, 188, 212, 0.2);
+}
