+++ css/login.css
@@ -1,0 +1,43 @@
+body {
+  background-color: #e0f7fa;
+  font-family: "Poppins", sans-serif;
+}
+
+.login-form {
+  background: #ffffff;
+  padding: 2.5rem;
+  border-radius: 16px;
+  box-shadow: 0 8px 20px rgba(0, 172, 193, 0.2);
+  max-width: 400px;
+  width: 100%;
+}
+
+h3 {
+  color: #00acc1;
+  font-weight: 600;
+}
+
+.btn-primary {
+  background-color: #00acc1;
+  border-color: #00acc1;
+  font-weight: 500;
+}
+
+.btn-primary:hover {
+  background-color: #00838f;
+  border-color: #00838f;
+}
+
+.form-control:focus {
+  border-color: #00acc1;
+  box-shadow: 0 0 0 0.25rem rgba(0, 172, 193, 0.2);
+}
+
+a {
+  color: #00acc1;
+  text-decoration: none;
+}
+
+a:hover {
+  text-decoration: underline;
+}
