+++ index.html
@@ -23,13 +23,12 @@
         </button>
 
         <div class="collapse navbar-collapse justify-content-end" id="navbarContent">
-          <!-- Guest view -->
+        
           <div id="guestLinks" class="d-flex gap-2">
             <a href="login.html" class="btn btn-outline-primary">Đăng nhập</a>
             <a href="register.html" class="btn btn-primary">Đăng ký</a>
           </div>
 
-          <!-- Logged-in view -->
           <div id="userDropdown" class="dropdown d-none">
             <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton"data-bs-toggle="dropdown" aria-expanded="false">
               <span id="userName">Người dùng</span>
@@ -67,9 +66,9 @@
           <div id="city" class="mb-2">--</div>
           <div id="today" class="text-muted mb-1">--</div>
           <div id="details">Vui lòng nhập tên thành phố để xem thời tiết.</div>
-          <div class="mt-3 text-center">
-            <button id="saveLocationBtn" class="btn btn-outline-warning d-none">🌟 Lưu địa điểm</button>
-          </div>
+            <div class="mt-3 text-center">
+              <button id="saveLocationBtn" class="btn btn-outline-warning d-none">🌟 Lưu địa điểm</button>
+            </div>
 
         </div>
 
@@ -101,6 +100,6 @@
     <script src="js/UI.js"></script>
     <script type="module" src="js/luudiadiem.js"></script>
     <script type="module" src="js/auth.js"></script>
-
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
