+++ login.html
@@ -3,57 +3,12 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Đăng nhập - Lavender Weather</title>
+  <title>Đăng nhập</title>
 
   <!-- Bootstrap + Google Font -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
-
-  <style>
-    body {
-      background-color: #e0f7fa;
-      font-family: 'Poppins', sans-serif;
-    }
-
-    .login-form {
-      background: #ffffff;
-      padding: 2.5rem;
-      border-radius: 16px;
-      box-shadow: 0 8px 20px rgba(0, 172, 193, 0.2);
-      max-width: 400px;
-      width: 100%;
-    }
-
-    h3 {
-      color: #00acc1;
-      font-weight: 600;
-    }
-
-    .btn-primary {
-      background-color: #00acc1;
-      border-color: #00acc1;
-      font-weight: 500;
-    }
-
-    .btn-primary:hover {
-      background-color: #00838f;
-      border-color: #00838f;
-    }
-
-    .form-control:focus {
-      border-color: #00acc1;
-      box-shadow: 0 0 0 0.25rem rgba(0, 172, 193, 0.2);
-    }
-
-    a {
-      color: #00acc1;
-      text-decoration: none;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-  </style>
+  <link rel="stylesheet" href="css/login.css">
 </head>
 
 <body>
@@ -86,6 +41,7 @@
   </div>
 
   <!-- Script -->
+  <script type="module" src="js/main.js"></script>
   <script type="module" src="js/auth.js"></script>
 </body>
 </html>
