+++ register.html
@@ -3,57 +3,11 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Đăng ký - Lavender Weather</title>
+  <title>Đăng ký</title>
 
-  <!-- Bootstrap + Google Font -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
-
-  <style>
-    body {
-      background: linear-gradient(to right, #e0f7fa, #ffffff);
-      font-family: 'Poppins', sans-serif;
-    }
-
-    .register-form {
-      background: #ffffff;
-      padding: 2.5rem;
-      border-radius: 16px;
-      box-shadow: 0 8px 20px rgba(0, 172, 193, 0.2);
-      max-width: 400px;
-      width: 100%;
-    }
-
-    h3 {
-      color: #00acc1;
-      font-weight: 600;
-    }
-
-    .form-control::placeholder {
-      font-style: italic;
-      color: #bbb;
-    }
-
-    .btn-primary {
-      background-color: #00acc1;
-      border-color: #00acc1;
-      font-weight: 500;
-    }
-
-    .btn-primary:hover {
-      background-color: #00838f;
-      border-color: #00838f;
-    }
-
-    a {
-      color: #00acc1;
-      text-decoration: none;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-  </style>
+  <link rel="stylesheet" href="css/login.css">
 </head>
 
 <body>
@@ -88,6 +42,6 @@
   </div>
 
   <!-- Script -->
-  <script type="module" src="js/auth.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
