--- t.htm
@@ -1,149 +1,0 @@
-<!DOCTYPE html>
-<html lang="vi">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Dự báo thời tiết</title>
-
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
-    <!-- Google Font -->
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
-
-    <!-- Custom CSS -->
-    <style>
-      body {
-        font-family: 'Poppins', sans-serif;
-        background-color: #e0f7fa;
-        color: #333;
-      }
-
-      .navbar {
-        background-color: #b2ebf2 !important;
-      }
-
-      .weather-card {
-        background-color: #ffffff;
-        border-radius: 12px;
-        padding: 20px;
-        margin-bottom: 24px;
-        box-shadow: 0 4px 10px rgba(0, 0, 0, 0.05);
-      }
-
-      .weather-card h5 {
-        font-weight: 600;
-      }
-
-      .btn-primary {
-        background-color: #00acc1;
-        border-color: #00acc1;
-      }
-
-      .btn-primary:hover {
-        background-color: #00838f;
-        border-color: #00838f;
-      }
-
-      #city {
-        font-size: 1.8rem;
-        font-weight: bold;
-        color: #007c91;
-      }
-
-      footer {
-        background-color: #b2ebf2;
-      }
-    </style>
-  </head>
-
-  <body class="d-flex flex-column min-vh-100">
-    <!-- Navbar -->
-    <nav class="navbar navbar-expand-lg navbar-light border-bottom">
-      <div class="container">
-        <a class="navbar-brand fw-bold text-primary" href="#">🌤️ Thời tiết</a>
-        <button class="navbar-toggler" data-bs-toggle="collapse" data-bs-target="#navbarContent"
-          aria-controls="navbarContent" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse justify-content-end" id="navbarContent">
-          <!-- Guest view -->
-          <div id="guestLinks" class="d-flex gap-2">
-            <a href="login.html" class="btn btn-outline-primary">Đăng nhập</a>
-            <a href="register.html" class="btn btn-primary">Đăng ký</a>
-          </div>
-
-          <!-- Logged-in view -->
-          <div id="userDropdown" class="dropdown d-none">
-            <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton"
-              data-bs-toggle="dropdown" aria-expanded="false">
-              <span id="userName">Người dùng</span>
-            </button>
-            <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="dropdownMenuButton">
-              <li><a class="dropdown-item" href="#">Tài khoản của tôi</a></li>
-              <li><hr class="dropdown-divider" /></li>
-              <li><a id="logout" class="dropdown-item" href="#">Đăng xuất</a></li>
-            </ul>
-          </div>
-        </div>
-      </div>
-    </nav>
-
-    <!-- Main -->
-    <main class="flex-fill">
-      <div class="container py-5">
-        <!-- Search -->
-        <div class="row justify-content-center mb-4">
-          <div class="col-md-10">
-            <div class="input-group shadow-sm">
-              <input type="text" id="timkiem" class="form-control" placeholder="Nhập tên thành phố..." />
-              <button class="btn btn-primary" onclick="TimKiem()">Tìm kiếm</button>
-              <button class="btn btn-outline-secondary" onclick="doiDonVi()">°C / °F</button>
-            </div>
-            <div id="error" class="text-danger mt-2"></div>
-          </div>
-        </div>
-
-        <!-- Saved locations -->
-        <div class="row justify-content-center mb-4">
-          <div class="col-md-10">
-            <h6>🌟 Địa điểm đã lưu:</h6>
-            <div id="savedLocations" class="d-flex gap-2 flex-wrap"></div>
-          </div>
-        </div>
-
-        <!-- Current weather -->
-        <div class="weather-card text-center">
-          <div id="city" class="mb-2">--</div>
-          <div id="today" class="text-muted mb-1">--</div>
-          <div id="details">Vui lòng nhập tên thành phố để xem thời tiết.</div>
-        </div>
-
-        <!-- Hourly Forecast -->
-        <div class="weather-card">
-          <h5 class="mb-3">⏰ Dự báo theo giờ</h5>
-          <div id="hourly-forecast"></div>
-        </div>
-
-        <!-- Daily Forecast -->
-        <div class="weather-card">
-          <h5 class="mb-3">📅 Dự báo theo ngày</h5>
-          <div id="daily-forecast"></div>
-        </div>
-      </div>
-    </main>
-
-    <!-- Footer -->
-    <footer class="text-center py-3 mt-auto border-top">
-      <div class="container">
-        <p class="mb-1">&copy; 2025 Dự báo thời tiết. Mọi quyền được bảo lưu.</p>
-      </div>
-    </footer>
-
-    <!-- Scripts -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="js/script.js"></script>
-    <script type="module" src="js/auth.js"></script>
-  </body>
-</html>
