+++ taikhoan.html
@@ -3,67 +3,12 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Tài khoản của tôi - Lavender Weather</title>
+  <title>Tài khoản của tôi</title>
 
-  <!-- Bootstrap + Google Font -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
 
-  <style>
-    body {
-      background-color: #e0f7fa;
-      font-family: 'Poppins', sans-serif;
-    }
-
-    .navbar {
-      background-color: #b2ebf2 !important;
-    }
-
-    .card-profile {
-      background-color: #ffffff;
-      border-radius: 16px;
-      padding: 2.5rem;
-      max-width: 600px;
-      margin: 60px auto;
-      box-shadow: 0 8px 20px rgba(0, 172, 193, 0.1);
-    }
-
-    h4 {
-      color: #00acc1;
-      font-weight: 600;
-      text-align: center;
-      margin-bottom: 1.5rem;
-    }
-
-    .form-control {
-      border-radius: 10px;
-    }
-
-    .form-label {
-      font-weight: 500;
-    }
-
-    .btn-save {
-      background-color: #00acc1;
-      border: none;
-      color: #fff;
-      font-weight: 500;
-      border-radius: 10px;
-    }
-
-    .btn-save:hover {
-      background-color: #00838f;
-    }
-
-    #previewAvatar {
-      width: 400px;
-      height: 400px;
-      border-radius: 50%;
-      border: 3px solid #00bcd4;
-      object-fit: cover;
-      box-shadow: 0 4px 8px rgba(0, 188, 212, 0.2);
-    }
-  </style>
+  <link rel="stylesheet" href="css/account.css" />
 </head>
 
 <body class="d-flex flex-column min-vh-100">
